build(separe the fkernel building into two steps): now we have a prekernel to boot the "kernel"

+++ Include/Kernel/Arch/x86_64/arch_defs.h
@@ -5,7 +5,7 @@
 /**
  * @brief Maximum number of interrupt vectors for x86_64
  */
-static constexpr uint8_t MAX_X86_64_INTERRUPTS_LENGTH = 256;
+static constexpr uint16_t MAX_X86_64_INTERRUPTS_LENGTH = 256;
 
 /**
  * @brief Maximum number of memory chunks per range
